+++ src/__tests__/lazy.test.js
@@ -22,10 +22,10 @@ describe('lazy', () => {
     });
 
     it('build self-referential types with lazy()', () => {
-        type StringList = {
+        type StringList = {|
             curr: string,
             next?: StringList,
-        };
+        |};
 
         const llist: Decoder<StringList> = object({
             curr: string,
@@ -44,10 +44,10 @@ describe('lazy', () => {
     });
 
     it('build self-referential types with variables', () => {
-        type Tree<T> = {
+        type Tree<T> = {|
             node: T,
             children: Array<Tree<T>>,
-        };
+        |};
 
         function tree<T>(decoder: Decoder<T>): Decoder<Tree<T>> {
             return object({