+++ src/object.js
@@ -5,13 +5,10 @@ import type { Annotation } from 'debrief';
 import { Err, Ok } from 'lemons/Result';
 
 import type { $DecoderType, Decoder } from './types';
-import { compose } from './utils';
+import { compose, map } from './utils';
 
 // $FlowFixMe (not really an issue) - deliberate use of `any` - not sure how we should get rid of this
 type AnyDecoder = any;
-
-// $FlowFixMe (not really an issue) - deliberately casting
-type cast = any;
 
 function isPojo(o: mixed): boolean %checks {
     return (
@@ -77,7 +74,7 @@ export const pojo: Decoder<{| [string]: mixed |}> = (blob: mixed) => {
  */
 export function object<O: { +[field: string]: AnyDecoder, ... }>(
     mapping: O
-): Decoder<$ObjMap<O, $DecoderType>> {
+): Decoder<$ObjMap<$Exact<O>, $DecoderType>> {
     const known = new Set(Object.keys(mapping));
     return compose(pojo, (blob: {| [string]: mixed |}) => {
         const actual = new Set(Object.keys(blob));
@@ -158,7 +155,7 @@ export function exact<O: { +[field: string]: AnyDecoder, ... }>(
 ): Decoder<$ObjMap<$Exact<O>, $DecoderType>> {
     // Check the inputted object for any superfluous keys
     const allowed = new Set(Object.keys(mapping));
-    const checked = compose(pojo, (blob: { [string]: mixed }) => {
+    const checked = compose(pojo, (blob: {| [string]: mixed |}) => {
         const actual = new Set(Object.keys(blob));
         const superfluous = subtract(actual, allowed);
         if (superfluous.size > 0) {
@@ -169,9 +166,28 @@ export function exact<O: { +[field: string]: AnyDecoder, ... }>(
         return Ok(blob);
     });
 
-    // Defer to the "object" decoder for doing the real decoding work.  Since
-    // we made sure there are no superfluous keys in this structure, it's now
-    // safe to force-cast it to an $Exact<> type.
-    const decoder = ((object(mapping): cast): Decoder<$ObjMap<$Exact<O>, $DecoderType>>);
-    return compose(checked, decoder);
+    return compose(checked, object(mapping));
+}
+
+export function inexact<O: { +[field: string]: AnyDecoder }>(
+    mapping: O
+): Decoder<$ObjMap<O, $DecoderType> & { +[string]: mixed }> {
+    return compose(pojo, (blob: {| [string]: mixed |}) => {
+        const allkeys = new Set(Object.keys(blob));
+        const decoder = map(object(mapping), (safepart: $ObjMap<O, $DecoderType>) => {
+            const safekeys = new Set(Object.keys(safepart));
+
+            // To account for hard-coded keys that aren't part of the input
+            for (const k of safekeys) {
+                allkeys.add(k);
+            }
+
+            const rv = {};
+            for (const k of allkeys) {
+                rv[k] = safekeys.has(k) ? safepart[k] : blob[k];
+            }
+            return rv;
+        });
+        return decoder(blob);
+    });
 }