+++ src/types/index.d.ts
@@ -24,7 +24,7 @@ export { instanceOf } from './instanceOf';
 export { lazy } from './lazy';
 export { mapping, dict } from './mapping';
 export { integer, number, positiveInteger, positiveNumber } from './number';
-export { exact, object, pojo } from './object';
+export { exact, inexact, object, pojo } from './object';
 export { maybe, nullable, optional } from './optional';
 export { email, regex, string, url } from './string';
 export { tuple2, tuple3, tuple4, tuple5, tuple6 } from './tuple';