+++ src/types/object-tests.ts
@@ -1,4 +1,4 @@
-import { exact, object, optional, pojo, string } from 'decoders';
+import { exact, inexact, object, optional, pojo, string } from 'decoders';
 
 // $ExpectType Decoder<{ bar: { qux: string; }; foo?: string | undefined; }, unknown>
 object({
@@ -11,6 +11,9 @@ exact({
     foo: optional(string),
     bar: object({ qux: string }),
 });
+
+// $ExpectType Decoder<{ id: string; } & { [extra: string]: unknown; }, unknown>
+inexact({ id: string });
 
 // $ExpectType Decoder<{ [key: string]: unknown; }, unknown>
 pojo;